# openapi/collaborative-translate-api.yaml
openapi: 3.0.0
paths:
  /project:
    get:
      operationId: getProjectsByQuery
      parameters:
        - name: query
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
    post:
      operationId: insertProject
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostProjectDto"
      responses:
        "201":
          description: ""
  "/project/{id}":
    get:
      operationId: getProjectById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    post:
      operationId: postProjectById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeProjectDto"
      responses:
        "201":
          description: ""
  "/project/{id}/files":
    get:
      operationId: getFiles
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    post:
      operationId: uploadFile
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
  "/project/{id}/language":
    get:
      operationId: getProjectLanguages
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    post:
      operationId: createTranslation
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostTranslateLanguage"
      responses:
        "201":
          description: ""
  "/project/{id}/actions":
    get:
      operationId: getAllActions
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  "/translation/{id}/suggestions":
    get:
      operationId: getSuggestions
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  "/translation/{id}":
    get:
      operationId: getTextPieceById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: nextMinLength
          required: true
          in: query
          schema:
            type: number
        - name: prevMinLength
          required: true
          in: query
          schema:
            type: number
        - name: toLanguageId
          required: true
          in: query
          schema:
            type: number
        - name: withOriginal
          required: true
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: ""
  "/translation/{id}/actions":
    get:
      operationId: getActions
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  "/translation/{id}/languages":
    get:
      operationId: getLanguages
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  "/translation/{id}/language":
    get:
      operationId: getLanguage
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  "/translation/{id}/suggest":
    post:
      operationId: suggest
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
  /translation:
    post:
      operationId: putTranslations
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "201":
          description: ""
  "/language/{id}":
    get:
      operationId: getLanguageById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  "/language/{id}/translations-orders":
    get:
      operationId: getTranslationsByOrder
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: orders
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
  "/language/{id}/translations":
    get:
      operationId: getTranslationsByLanguage
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: fileId
          required: true
          in: query
          schema:
            type: string
        - name: take
          required: true
          in: query
          schema:
            type: number
        - name: page
          required: true
          in: query
          schema:
            type: number
        - name: shouldTranslate
          required: true
          in: query
          schema:
            type: boolean
        - name: withOriginal
          required: true
          in: query
          schema:
            type: boolean
        - name: status
          required: true
          in: query
          schema:
            type: string
        - name: hasSuggestions
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
  "/language/{id}/assemble":
    post:
      operationId: assembleFiles
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
  "/language/{id}/assemblies":
    get:
      operationId: getAssemblies
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  "/suggestions/{id}/approve":
    post:
      operationId: approve
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
  "/suggestions/{id}/deny":
    post:
      operationId: deny
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      security:
        - bearer: []
  "/files/{id}/peek":
    post:
      operationId: getPeek
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
  "/files/{id}/split":
    post:
      operationId: splitFile
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
  "/files/{id}/assemble":
    get:
      operationId: assembleFile
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: languageId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
  "/assembly/{id}":
    get:
      operationId: getAssemblyById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  /user/me:
    get:
      operationId: getProfile
      parameters: []
      responses:
        default:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserDto"
      security:
        - bearer: []
  "/user/{id}":
    get:
      operationId: getUserById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
    post:
      operationId: changeUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeUserDto"
      responses:
        "201":
          description: ""
  "/user/{id}/projects":
    get:
      operationId: getProjectsByUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
  /auth/login:
    post:
      operationId: login
      parameters: []
      responses:
        "201":
          description: ""
  /auth/register:
    post:
      operationId: register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUserDto"
      responses:
        "201":
          description: ""
  /auth/refresh:
    get:
      operationId: refresh
      parameters: []
      responses:
        "200":
          description: ""
  /regexp:
    get:
      operationId: getAll
      parameters: []
      responses:
        "200":
          description: ""
info:
  title: Collaborative Translate API
  description: Collaborative Translate web-app API specification
  version: "1.0"
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    cookie:
      type: apiKey
      in: cookie
      name: connect.sid
  schemas:
    ChangeProjectDto:
      type: object
      properties: {}
    PostProjectDto:
      type: object
      properties: {}
    PostTranslateLanguage:
      type: object
      properties: {}
    GetUserDto:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        info:
          type: string
      required:
        - id
        - name
        - email
        - info
    ChangeUserDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        info:
          type: string
    PostUserDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required:
        - email
        - password
        - name